requirements.txt
absl-py==1.2.0
addict==2.4.0
attrs==22.1.0
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.0.12
click==7.1.2
commonmark==0.9.1
cycler==0.11.0
fastapi==0.65.1
fonttools==4.33.3
google-auth==2.9.1
google-auth-oauthlib==0.4.6
grpcio==1.47.0
h11==0.14.0
idna==3.3
importlib-metadata==4.12.0
kiwisolver==1.4.3
Markdown==3.4.1
matplotlib==3.5.2
mediapipe==0.8.10.1
numpy==1.21.6
oauthlib==3.2.0
onnx==1.12.0
onnx-simplifier==0.4.1
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
packaging==21.3
pandas==1.3.5
Pillow==9.1.1
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.2
Pygments==2.12.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2022.1
PyYAML==6.0
requests==2.28.0
requests-oauthlib==1.3.1
rich==12.5.1
rsa==4.9
scipy==1.7.3
six==1.16.0
starlette==0.14.2
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
thop==0.1.1.post2207130030
torch==1.11.0
torchaudio==0.11.0
torchvision==0.12.0
tqdm==4.64.0
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.13.4
Werkzeug==2.1.2
zipp==3.8.1
